Main Page Update
Updated the Main Page code to accept dynamic text input and profile
image.  Please see attachment for the code changes.  I have placed them
in its own group titled "Main Page"

For the three sections of the Main Page - Profile, Upcoming Event,
Group Feed please see the following notes.

Profile:
Data contents are defined in Profile object, currently there is only
one instance of this object and it is hardcoded but it can be expanded
to use the json to populate its members

Upcoming Event:
Data is contained in the Event object and the Main window tracks a list
of Events.
However, currently only the first object is displayed on the screen.
This is also hardcoded but can be modified with the json for population.

Group Feed:
The title for the section of group feed is defined in the Profile
object under instance Group.
I changed the Feed to be a text box. This can be changed later to multi
cell or something else to better fit the data.

This is also hardcoded but can be modified with the json for population.