Group name and Main Page Database
Fixed the page segue flow for login to main page.

Worked on the local database and fixed issues with sharing data.  There
is an ActiveProfile which is a singleton which can be used past the
login page.

GroupField can now be updated.

Main Page first section now uses the database.